requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pyodbc==5.2.0
PySocks==1.7.1
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
selenium==4.34.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0

Faker~=37.4.0

